Use weak instead of unowned in all closures; Use values in range of 0..1 for color components instead of 0..255

+++ Sources/LifeServer/UsersManager.swift
@@ -18,9 +18,9 @@
 import Foundation
 
 public struct Color : Codable {
-    var r: Int = 0
-    var g: Int = 0
-    var b: Int = 0
+    var r: Double = 0
+    var g: Double = 0
+    var b: Double = 0
 }
 
 public struct User : Codable {
